update traditional models and add comments on data_cleaning

+++ waifu_project/data_cleaning.py
@@ -4,45 +4,58 @@ import shutil
 
 
 def data_cleaning():
-    source_path = os.path.join('data_set', 'self-collected')  # 'data_set/self-collected (copy)'
+    # set source path and destination path
+    source_path = os.path.join('data_set', 'self-collected')
     destination_path = os.path.join('data_set', 'modeling_data')
     source_pics_folders = os.listdir(source_path)
 
+    # find out doujin pictures and official pictures
     doujin_re = '.*doujin.*'
     official_re = '.*official.*'
     doujin = [x for x in source_pics_folders if re.match(doujin_re, x)]
     official = [x for x in source_pics_folders if re.match(official_re, x)]
 
-    shutil.rmtree(destination_path)
+    # clean the destination path
+    if os.path.exists(destination_path):
+        shutil.rmtree(destination_path)
     os.makedirs(destination_path)
 
+
+    # copy official pics to destination folder
     for file in official:
+        # find the picture folder
         pic_folder_path = os.path.join(source_path, file)
-        modified_file_name = file[6:-9]
-        destination_folder = os.path.join(destination_path, modified_file_name)
+        # extract character's name from folder's name
+        modified_folder_name = file[6:-9]
+        destination_folder = os.path.join(destination_path, modified_folder_name)
+        # list all the pictures under the folder
         pics = os.listdir(pic_folder_path)
         if not os.path.exists(destination_folder):
             os.makedirs(destination_folder)
-        print("COPYING PICS IN {} TO {}".format(pic_folder_path, destination_folder))
+        print("[STATUS] copying pics in {} to {}".format(pic_folder_path, destination_folder))
+        # copy
         for pic in pics:
-            # if re.match(official_re, pic) is None:
             shutil.copy(os.path.join(pic_folder_path, pic), destination_folder)
             os.rename(os.path.join(destination_folder, pic), os.path.join(destination_folder, "official_" + pic))
-        print("{} COPY WORK FINISHED".format(file))
+        print("[STATUS] {} copy work finished".format(file))
 
+    # copy doujin pics to destination folder
     for file in doujin:
+        # find the picture folder
         pic_folder_path = os.path.join(source_path, file)
-        modified_file_name = file[6:-7]
-        destination_folder = os.path.join(destination_path, modified_file_name)
-        print("COPYING PICS IN {} TO {}".format(pic_folder_path, destination_folder))
+        # extract character's name from folder's name
+        modified_folder_name = file[6:-7]
+        destination_folder = os.path.join(destination_path, modified_folder_name)
+        print("[STATUS] copying pics in {} to {}".format(pic_folder_path, destination_folder))
+        # list all the pictures under the folder
         pics = os.listdir(pic_folder_path)
         if not os.path.exists(destination_folder):
             os.makedirs(destination_folder)
+        # copy
         for pic in pics:
-            # if re.match(doujin_re, pic) is None:
             shutil.copy(os.path.join(pic_folder_path, pic), destination_folder)
             os.rename(os.path.join(destination_folder, pic), os.path.join(destination_folder, "doujin_" + pic))
-        print("{} COPY WORK FINISHED".format(file))
+        print("[STATUS] {} copy work finished".format(file))
 
 
 if __name__ == '__main__':
